Add extra map for carrying along information while encoding/decoding
The headers in a DICOM file is unfortunately not always know before hand.
Especially the pixel data header might be OB or OB.
When the data is explicitly encoded it is easy, however, for implicitly
encoded/decoded data it becomes trickier. Because of this, we need
to use the "Extra" map. Where we accumulate information about the
dataset while we are encoding/decoding.