data_element: Short-cut for empty encoding
There are some datasets that have empty headers.
This is something that we don't really expected.

E.g, wolfpacs needs to be able to encode a US that is <<>>.
This does not really make sense but some applications do this.

Therefore we adopt the simple rule. <<>> is encoded ass <<>>,
no matter the VR.