wolfpacs_workers: Add feature to pause worker
It is sometimes important to take a worker offline in
controlled manner. Therefore, add support to pause a worker.
That way, the worker will continue to receive series for all
mapped studies, but not receive any new studies.

Once all the studies that are mapped are finished (which might take
many minutes), the worker is safe to take offline. When the worker
is ready for new jobs again - just simply unpause it.